requirements.txt
Flask
pytesseract
PyMuPDF
Pillow
nltk
transformers
torch